+++ backend/danswer/configs/constants.py
@@ -14,7 +14,6 @@ ALLOWED_USERS = "allowed_users"
 ALLOWED_GROUPS = "allowed_groups"
 METADATA = "metadata"
 GEN_AI_API_KEY_STORAGE_KEY = "genai_api_key"
-HTML_SEPARATOR = "\n"
 PUBLIC_DOC_PAT = "PUBLIC"
 QUOTE = "quote"
 BOOST = "boost"
