+++ backend/danswer/connectors/confluence/connector.py
@@ -1,3 +1,4 @@
+import os
 from collections.abc import Callable
 from collections.abc import Collection
 from datetime import datetime
@@ -101,12 +102,43 @@ class ConfluenceConnector(LoadConnector, PollConnector):
         start_ind: int,
     ) -> Collection[dict[str, Any]]:
         def _fetch(start_ind: int, batch_size: int) -> Collection[dict[str, Any]]:
-            return confluence_client.get_all_pages_from_space(
-                self.space,
-                start=start_ind,
-                limit=batch_size,
-                expand="body.storage.value,version",
-            )
+            try:
+                return confluence_client.get_all_pages_from_space(
+                    self.space,
+                    start=start_ind,
+                    limit=batch_size,
+                    expand="body.storage.value,version",
+                )
+            except:
+                logger.warning(
+                    f"Batch failed with space {self.space} at offset {start_ind}"
+                )
+
+                view_pages: list[dict[str, Any]] = []
+                for i in range(self.batch_size):
+                    try:
+                        # Could be that one of the pages here failed due to this bug:
+                        # https://jira.atlassian.com/browse/CONFCLOUD-76433
+                        view_pages.extend(
+                            confluence_client.get_all_pages_from_space(
+                                self.space,
+                                start=start_ind + i,
+                                limit=1,
+                                expand="body.storage.value,version",
+                            )
+                        )
+                    except:
+                        # Use view instead, which captures most info but is less complete
+                        view_pages.extend(
+                            confluence_client.get_all_pages_from_space(
+                                self.space,
+                                start=start_ind + i,
+                                limit=1,
+                                expand="body.view.value,version",
+                            )
+                        )
+
+                return view_pages
 
         try:
             return _fetch(start_ind, self.batch_size)
@@ -162,7 +194,10 @@ class ConfluenceConnector(LoadConnector, PollConnector):
             last_modified = datetime.fromisoformat(last_modified_str)
 
             if time_filter is None or time_filter(last_modified):
-                page_html = page["body"]["storage"]["value"]
+                page_html = (
+                    page["body"].get("storage", {}).get("value")
+                    or page["body"]["view"]["value"]
+                )
                 page_text = (
                     page.get("title", "") + "\n" + parse_html_page_basic(page_html)
                 )
@@ -219,3 +254,15 @@ class ConfluenceConnector(LoadConnector, PollConnector):
 
             if num_pages < self.batch_size:
                 break
+
+
+if __name__ == "__main__":
+    connector = ConfluenceConnector(os.environ["CONFLUENCE_TEST_SPACE_URL"])
+    connector.load_credentials(
+        {
+            "confluence_username": os.environ["CONFLUENCE_USER_NAME"],
+            "confluence_access_token": os.environ["CONFLUENCE_ACCESS_TOKEN"],
+        }
+    )
+    document_batches = connector.load_from_state()
+    print(next(document_batches))
